Move ui_screen to the screen module
Submodules for this module will be handlers and helpers inherited by
UIScreen.
