+++ examples/hub_spoke/hub_spoke.py
@@ -1,10 +1,10 @@
 #!/bin/python3
 
 from simpleline.base import App
-from simpleline.render.screen_scheduler import INPUT_PROCESSED, INPUT_DISCARDED
 from simpleline.render.adv_widgets import PasswordDialog
 from simpleline.render.prompt import Prompt
-from simpleline.render.ui_screen import UIScreen
+from simpleline.render.screen import UIScreen
+from simpleline.render.screen_scheduler import INPUT_PROCESSED, INPUT_DISCARDED
 from simpleline.render.widgets import TextWidget, ColumnWidget, CenterWidget
