+++ tests/screen_scheduler_test.py
@@ -18,10 +18,11 @@
 #
 
 import unittest
-from unittest import mock
 from io import StringIO
+from unittest import mock
+
 from simpleline.base import App
-from simpleline.render.ui_screen import UIScreen
+from simpleline.render.screen import UIScreen
 from tests import schedule_screen_and_run, create_output_with_separators
