+++ tests/signal_handler_test.py
@@ -21,10 +21,9 @@
 import unittest
 from unittest.mock import patch, MagicMock
 
-from simpleline.render.ui_screen import UIScreen
-from simpleline.event_loop.signals import RenderScreenSignal, AbstractSignal
-
 from simpleline.base import App
+from simpleline.event_loop.signals import RenderScreenSignal, AbstractSignal
+from simpleline.render.screen import UIScreen
 
 
 class SignalHandler_TestCase(unittest.TestCase):
